backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Database
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
bcrypt==4.3.0

# Validation & settings
pydantic==2.11.1
pydantic-settings==2.8.1
email-validator==2.2.0

# Logging
structlog==25.1.0

# ASGI
httptools==0.6.4

# Testing
httpx==0.28.1
pytest==8.3.5
anyio==4.13.0
pytest-asyncio==0.26.0
